requirements.txt
alembic==0.9.9
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.2.2
decorator==4.3.0
falcon==1.4.1
gunicorn==19.8.1
idna==2.6
ipython==6.4.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.0
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.2
packaging==17.1
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
pip-review==1.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.1
pymodm==0.4.0
pymongo==3.6.1
pyparsing==2.2.0
python-dateutil==2.7.3
python-editor==1.0.3
python-mimeparse==1.6.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.7
traitlets==4.3.2
ujson==1.35
wcwidth==0.1.7
